app/requirements.txt
Flask==2.3.3
torchvision
gensim
matplotlib
numpy
nltk
scipy